+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Changed
 
 - Changed `RedirectedProcess` to have Builder-like API.
+- Improved the UX of the BuildConfiguration inspector.
+- Improved the UX of the GDK Tools Configuration window.
+- Changed the format of the BuildConfiguration asset. Please recreate, or copy it from `workers/unity/Playground/Assets/Config/BuildConfiguration.asset`.
 
 ### Fixed